Adds Streamed spec (WIP)
